Move static optimizie plugin code to a separate dir, update README
