+++ src/static-build/StaticOptimizePlugin.ts
