+++ src/static-build/features/chrome.json
