+++ src/static-build/features/firefox.json
