+++ src/static-build/features/safari.json
