+++ src/static-build/util/walk.ts
