+++ src/static-build/webpack.d.ts
