+++ tests/unit/StaticOptimizePlugin.ts
@@ -7,7 +7,7 @@ import Compilation = require('../support/webpack/Compilation');
 import Compiler = require('../support/webpack/Compiler');
 import ConstDependency = require('webpack/lib/dependencies/ConstDependency');
 import NormalModule = require('webpack/lib/NormalModule');
-import _StaticOptimizePlugin from '../../src/StaticOptimizePlugin';
+import _StaticOptimizePlugin from '../../src/static-build/StaticOptimizePlugin';
 
 function getAst(name: string): Program {
 	return (require as any).nodeRequire((require as any).toUrl(`../support/fixtures/${name}.json`));
